Add logic to display Loading/Empty or Characters
The redux store saves each successful search as an array. Therefore if
characters is falsy, it means the server hasn't responded yet, which
would at minimum be an empty array.

If an empty array is returned, the Empty component should be returned,
with a custom message based on whether the search field is currently
empty or not, indicated by the isSearching prop.