Add bookmarks and isSearching prop to CharacterListContainer
bookmarks is an object, with all the bookmarks stored with their id as
a key, and the full character object as the value. This allows for fast
lookup times.

Since the CharacterList can only take an array of characters, the
object has to be transformed into an array before being passed as a
prop.